Apply Coding Standards to internal scripts
