+++ phpcs.xml
@@ -3,9 +3,7 @@
 	<description>WordPress Coding Standards for Plugins</description>
 
 	<!-- Exclude some directories. -->
-	<exclude-pattern>/.scripts/*</exclude-pattern>
 	<exclude-pattern>/.wordpress-org/*</exclude-pattern>
-	<exclude-pattern>/lib/browser.php</exclude-pattern>
 	<exclude-pattern>/tests/*</exclude-pattern>
 	<exclude-pattern>/vendor/*</exclude-pattern>
